Refactor project overview list columns
Two columns, the status column and the grades column, belong into their respective subsystem.
New folders `templatetags` are created there to store future tag additions.
